+++ noodles-bam/src/bai/index.rs
@@ -1,15 +1,16 @@
 //! BAM index and fields.
 
 mod builder;
-pub mod reference_sequence;
 
-pub use self::{builder::Builder, reference_sequence::ReferenceSequence};
+pub use self::builder::Builder;
 
 use std::io;
 
 use noodles_core::{region::Interval, Position};
 use noodles_csi::{
-    binning_index::optimize_chunks, index::reference_sequence::bin::Chunk, BinningIndex,
+    binning_index::optimize_chunks,
+    index::{reference_sequence::bin::Chunk, ReferenceSequence},
+    BinningIndex,
 };
 
 const MIN_SHIFT: u8 = 14;
@@ -109,7 +110,7 @@ impl BinningIndex for Index {
             })?;
 
         let query_bins = reference_sequence
-            .query(interval)
+            .query(MIN_SHIFT, DEPTH, interval)
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
 
         let chunks: Vec<_> = query_bins
@@ -119,7 +120,7 @@ impl BinningIndex for Index {
             .collect();
 
         let (start, _) = resolve_interval(interval)?;
-        let min_offset = reference_sequence.min_offset(start);
+        let min_offset = reference_sequence.min_offset(MIN_SHIFT, DEPTH, start);
         let merged_chunks = optimize_chunks(&chunks, min_offset);
 
         Ok(merged_chunks)